requirements.txt
flask
flask-sqlalchemy
Flask-DB
Flask-Migrate